refactor: API 명세 변경에 따른 패키지 구조 변경, 체험단-활동 로그 페이지 프로토타입 개발
