+++ src/main/java/com/nuguna/freview/dto/cust/brand/CustMyZzimStoreDto.java
@@ -1,0 +1,12 @@
+package com.nuguna.freview.dto.cust.brand;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class CustMyZzimStoreDto {
+
+}
