+++ src/main/java/com/nuguna/freview/servlet/member/page/customer/my/ActivityPageServlet.java
@@ -1,0 +1,26 @@
+package com.nuguna.freview.servlet.member.page.customer.my;
+
+import java.io.IOException;
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+@WebServlet("/my-activity")
+public class ActivityPageServlet extends HttpServlet {
+
+  @Override
+  public void init() throws ServletException {
+  }
+
+  @Override
+  public void service(HttpServletRequest request, HttpServletResponse response)
+      throws ServletException, IOException {
+
+  }
+
+  @Override
+  public void destroy() {
+  }
+}
